hub/requirements.txt
Django==2.2.1
django-appconf==1.0.3
django-filter==2.1.0
django-imagekit==4.0.2
djangorestframework==3.9.4
pilkit==2.0
Pillow==6.0.0
pkg-resources==0.0.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
